improve tablewriter error handling
If opening a tablewriter Column hits the OS ulimit `OSError: [Errno 24] Too many open files`, the Column's `close()` method needn't compound the stack traceback by trying to access `self._data` which didn't get set.

Did WCM add more tables? I added `ulimit -n 4096` to my Mac laptop's bash profile, overriding the default of 256. Add this step to environment setup instructions?

+++ wholecell/io/tablewriter.py
@@ -238,12 +238,13 @@ class _Column(object):
 		-----
 		Trying to append after closing will raise an error.
 		"""
-		if not self._data.closed:
-			try:
-				self._write_block()
-				self._data.truncate()
-			finally:
-				self._data.close()
+		if hasattr(self, '_data'):
+			if not self._data.closed:
+				try:
+					self._write_block()
+					self._data.truncate()
+				finally:
+					self._data.close()
 
 
 	def __del__(self):
